move window size watch in custom hook
